テストの細かな改善
