+++ test/unlinkFunction.test.ts
@@ -1,7 +1,5 @@
-import { spawn } from 'node:child_process';
-import { existsSync, unlinkSync } from 'node:fs';
-import * as path from 'node:path';
-import { beforeEach, describe, expect, it, vi, type MockedFunction } from 'vitest';
+import { unlinkSync } from 'node:fs';
+import { afterEach, describe, expect, it, vi, type MockedFunction } from 'vitest';
 
 const mockUnlinkSync = vi.hoisted(() => vi.fn()) as MockedFunction<typeof unlinkSync>;
 const mockProcessOn = vi.hoisted(() => vi.fn());
@@ -20,11 +18,12 @@ vi.mock('node:process', () => ({
 }));
 
 // テスト実行前にモジュールをインポート
-import { prepareUnlinkFunction } from '../src/unlinkFunction';
+import { prepareUnlinkFunction, __resetUnlinkFunctionsForTest } from '../src/unlinkFunction';
 
 describe('prepareUnlinkFunction', () => {
-  beforeEach(() => {
+  afterEach(() => {
     vi.clearAllMocks();
+    __resetUnlinkFunctionsForTest();
   });
 
   it('should return a function that deletes the specified file', () => {
@@ -155,56 +154,4 @@ describe('prepareUnlinkFunction', () => {
 
     expect(() => prepareUnlinkFunction(filePath)).toThrow('filePath must not be empty');
   });
-
-  it('should delete files on process exit (integration test)', async () => {
-    const scriptPath = path.resolve(__dirname, 'scripts', 'unlinkOnExit.ts');
-
-    return new Promise<void>((resolve, reject) => {
-      let tempFilePath = '';
-      let stdout = '';
-
-      const child = spawn('npx', ['tsx', scriptPath], {
-        stdio: 'pipe'
-      });
-
-      child.stdout.on('data', (data) => {
-        stdout += data.toString();
-        const match = stdout.match(/TEMP_FILE_PATH:(.+)/);
-        if (match) {
-          tempFilePath = match[1].trim();
-
-          // ファイルが存在することを確認
-          expect(existsSync(tempFilePath)).toBe(true);
-        }
-      });
-
-      child.stderr.on('data', (data) => {
-        // vscode のターミナルから実行すると debugger 関連のメッセージが stderr に出力されるので無視する
-        if (/debugger/i.test(data.toString())) return;
-
-        reject(new Error(`Script stderr: ${data.toString()}`));
-      });
-
-      child.on('exit', (code) => {
-        try {
-          // プロセスが正常終了したことを確認
-          expect(code).toBe(0);
-
-          // 一時ファイルのパスが取得できていることを確認
-          expect(tempFilePath).toBeTruthy();
-
-          // プロセス終了後、ファイルが削除されていることを確認
-          expect(existsSync(tempFilePath)).toBe(false);
-
-          resolve();
-        } catch (error) {
-          reject(error);
-        }
-      });
-
-      child.on('error', (error) => {
-        reject(new Error(`Failed to spawn process: ${error.message}`));
-      });
-    });
-  }, 10000); // 10秒のタイムアウトを設定
 });
